requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.1
mss==9.0.1
Pillow==10.4.0
pynput==1.7.6
requests==2.32.3

